+++ src/core/parser/script.cpp
@@ -409,13 +409,13 @@ void spline(float x[], float y[], int n, float y2[])
 
 	for (i=2; i<=n-1; i++) {
 		sig = (x[i] - x[i-1])/(x[i+1] - x[i-1]);
-		p = sig * y2[i-1] + 2.0;
-		y2[i] = (sig - 1.0) / p;
+		p = sig * y2[i-1] + 2.0f;
+		y2[i] = (sig - 1.0f) / p;
 		u[i] = (y[i+1] - y[i])/(x[i+1] - x[i]) - (y[i] - y[i-1])/(x[i] - x[i-1]);
-		u[i] = (6.0*u[i]/(x[i+1] - x[i-1]) - sig*u[i-1])/p;
+		u[i] = (6.0f*u[i]/(x[i+1] - x[i-1]) - sig*u[i-1])/p;
 	}
-	qn=un=0.0;
-	y2[n]=(un - qn*u[n-1])/(qn * y2[n-1] + 1.0);
+	qn=un=0.0f;
+	y2[n]=(un - qn*u[n-1])/(qn * y2[n-1] + 1.0f);
 	for (k=n-1; k>=1; k--) {
 		y2[k] = y2[k] * y2[k+1] + u[k];
 	}
@@ -442,7 +442,7 @@ int splint(float xa[], float ya[], float y2a[], int n, float x, float * y, bool 
 	b = (x - xa[klo])/h;
 
 	if (cubic) {
-		*y = a * ya[klo] + b*ya[khi] + ((a*a*a - a)*y2a[klo] + (b*b*b - b)*y2a[khi]) * (h*h) / 6.0;
+		*y = a * ya[klo] + b*ya[khi] + ((a*a*a - a)*y2a[klo] + (b*b*b - b)*y2a[khi]) * (h*h) / 6.0f;
 	} else {
 		*y = a * ya[klo] + b*ya[khi];
 	}
@@ -482,7 +482,7 @@ AVSValue Spline(AVSValue args, void*, IScriptEnvironment* env )
 
 	AVSValue coordinates;
 
-	x = args[0].AsFloat(0);
+	x = (float)args[0].AsFloat(0);
 	coordinates = args[1];
 	cubic = args[2].AsBool(true);
 
@@ -492,8 +492,8 @@ AVSValue Spline(AVSValue args, void*, IScriptEnvironment* env )
 
 	n=n/2;
 	for (i=1; i<=n; i++) {
-		xa[i] = coordinates[(i-1)*2].AsFloat(0);
-		ya[i] = coordinates[(i-1)*2+1].AsFloat(0);
+		xa[i] = (float)coordinates[(i-1)*2].AsFloat(0);
+		ya[i] = (float)coordinates[(i-1)*2+1].AsFloat(0);
 	}
 
 	for (i=1; i<n; i++) {
